requirements.txt
easyocr==1.7.2
pytesseract==0.3.13
opencv-python==4.11.0.86
numpy==2.2.2
pandas==2.2.3
python-dotenv==1.0.1
pillow==10.4.0
streamlit==1.41.1
PyMuPDF==1.25.2
plotly==5.24.1
requests==2.32.3
together==1.3.13
cloudinary==1.42.1

